fixed output class and file structure
