+++ src/dp/course.py
@@ -1,5 +1,5 @@
 from array import *
-from ..utils.output import *
+from ..io.output import *
 import json
 
 import logging
