+++ src/dp/parse.py
@@ -1,4 +1,4 @@
-from ..utils.output import *
+from ..io.output import *
 from .course import Course
 from .course_template import Template
 from .catalog import Catalog
